docker-onnx-cpu/requirements.txt
pillow==10.0.0
SQLAlchemy==2.0.19
pgvector==0.2.1
PyYAML==6.0.1
loguru==0.7.0
fastapi==0.100.1
pydantic==2.1.1
psutil==5.9.5
uvicorn==0.23.2
psycopg2-binary==2.9.7
python-multipart==0.0.6
onnxruntime==1.15.1
six==1.16.0
